requirements.txt
numpy=1.18.1
matplotlib=3.1.3
pillow=4.0.0
